requirements.txt
gps3==0.33.3
python-socketio==4.3.1
requests==2.22.0